Refactored uiLanguages into own file in assorted. Moved icons to public
