+++ public/static/icons/dan-deselected.png
