+++ public/static/icons/dan-selected.png
