+++ src/assorted/uiLanguages.js
@@ -1,0 +1,14 @@
+export default [
+    {
+      name: "dansk",
+      code: "da",
+      selectedIcon: "/static/icons/dan-selected.png",
+      deselectedIcon: "/static/icons/dan-deselected.png",
+    },
+    {
+      name: "english",
+      code: "en",
+      selectedIcon: "/static/icons/eng-selected.png",
+      deselectedIcon: "/static/icons/eng-deselected.png",
+    },
+  ];
