+++ src/pages/Settings.js
@@ -1,7 +1,7 @@
 import { useEffect, useState, useContext } from "react";
 import { useHistory } from "react-router-dom";
 
-import { LanguageSelector } from "../components/LanguageSelector";
+import UiLanguageSelector from "../components/UiLanguageSelector";
 
 import { UserContext } from "../UserContext";
 import LoadingAnimation from "../components/LoadingAnimation";
@@ -11,6 +11,7 @@ import * as sc from "../components/TopTabs.sc";
 import { setTitle } from "../assorted/setTitle";
 
 import LocalStorage from "../assorted/LocalStorage";
+import uiLanguages from '../assorted/uiLanguages'
 
 import strings from "../i18n/definitions";
 
@@ -22,16 +23,7 @@ export default function Settings({ api, setUser }) {
   const [languages, setLanguages] = useState();
 
   // TODO: Refactor using Zeeguu project logic
-  const uiLanguages = [
-    {
-      name: "Dansk",
-      code: "da",
-    },
-    {
-      name: "English",
-      code: "en",
-    },
-  ];
+
 
   const [uiLanguage, setUiLanguage] = useState();
 
@@ -117,7 +109,7 @@ export default function Settings({ api, setUser }) {
         />
 
         <label>{strings.learnedLanguage}</label>
-        <LanguageSelector
+        <UiLanguageSelector
           languages={languages.learnable_languages}
           selected={language_for_id(
             userDetails.learned_language,
@@ -133,7 +125,7 @@ export default function Settings({ api, setUser }) {
         />
 
         <label>{strings.nativeLanguage}</label>
-        <LanguageSelector
+        <UiLanguageSelector
           languages={languages.native_languages}
           selected={language_for_id(
             userDetails.native_language,
@@ -144,7 +136,7 @@ export default function Settings({ api, setUser }) {
         {uiLanguage.name !== undefined && (
           <>
             <label>{strings.systemLanguage}</label>
-            <LanguageSelector
+            <UiLanguageSelector
               languages={uiLanguages}
               selected={uiLanguage.name}
               onChange={(e) => {
